Add sibling under the current node is considered only 1 commands and not 2
Before, when adding a sibling and pressing escape, the new node was moved to the last sibling position instead of being deleted. This was due to the fact that there was 2 commands : add sibling + move to correct position.
Now, the move position is not stored in history, then pressing escape indeed removes the node.